Support larger runs (thread IDs uint16 => uint32) (#428)
We hit the limit for thread cardinality in a large object storage
benchmark run (500 hosts, 192 concurrency).  This patch bumps the thread
ID data type from uint16 to uint32 to support larger runs.

+++ cli/merge.go
@@ -66,7 +66,7 @@ func mainMerge(ctx *cli.Context) error {
 	zstdDec, _ := zstd.NewReader(nil)
 	defer zstdDec.Close()
 	var allOps bench.Operations
-	threads := uint16(0)
+	threads := uint32(0)
 	log := console.Printf
 	if globalQuiet {
 		log = nil
