+++ pkg/bench/versioned.go
@@ -163,7 +163,7 @@ func (g *Versioned) Start(ctx context.Context, wait chan struct{}) error {
 					client, clDone := g.Client()
 					op := Operation{
 						OpType:   operation,
-						Thread:   uint16(i),
+						Thread:   uint32(i),
 						Size:     obj.Size,
 						File:     obj.Name,
 						ObjPerOp: 1,
@@ -203,7 +203,7 @@ func (g *Versioned) Start(ctx context.Context, wait chan struct{}) error {
 					client, clDone := g.Client()
 					op := Operation{
 						OpType:   operation,
-						Thread:   uint16(i),
+						Thread:   uint32(i),
 						Size:     obj.Size,
 						File:     obj.Name,
 						ObjPerOp: 1,
@@ -239,7 +239,7 @@ func (g *Versioned) Start(ctx context.Context, wait chan struct{}) error {
 					obj := g.Dist.deleteRandomObj()
 					op := Operation{
 						OpType:   operation,
-						Thread:   uint16(i),
+						Thread:   uint32(i),
 						Size:     0,
 						File:     obj.Name,
 						ObjPerOp: 1,
@@ -259,7 +259,7 @@ func (g *Versioned) Start(ctx context.Context, wait chan struct{}) error {
 					client, clDone := g.Client()
 					op := Operation{
 						OpType:   operation,
-						Thread:   uint16(i),
+						Thread:   uint32(i),
 						Size:     0,
 						File:     obj.Name,
 						ObjPerOp: 1,
